# Cached the calculation of the field's label and source attributes