--- a/rest_framework/compat.py
+++ b/rest_framework/compat.py
@@ -189,9 +189,9 @@ except ImportError:
 
     def _make_key(args, kwds, typed,
                  kwd_mark = (object(),),
-                 fasttypes = {int, str, frozenset, type(None)},
+                 fasttypes = set((int, str, frozenset, type(None))),
                  sorted=sorted, tuple=tuple, type=type, len=len):
-        'Make a cache key from optionally typed positional and keyword arguments'
+        """Make a cache key from optionally typed positional and keyword arguments"""
         key = args
         if kwds:
             sorted_items = sorted(kwds.items())