This results in much faster binding times and less initialization overhead even between different serializers who use the same source.