@tomchristie 2015-08-28
Closing as described.