@thedrow 2015-10-08
Since we dropped 1.4 support and all Django versions now include the lru_cache compatibility layer we should reconsider this PR.
@tomchristie what do you think? I can edit it to remove the branch that vendors it if it doesn't exist in Django.