@xordoquy 2015-10-08
@thedrow Do you have some numbers about how much it affects performances ?
We're not going to accept changes that make code harder to maintain just for the sake of it.