@thedrow 2015-10-08
Ok so let's review the changes I introduced here:
```python
return field_name.replace('_', ' ').capitalize()
```
performs two loops, one for replacing the string's characters and one to capitalize the string. Each comes with it's own memory allocations as opposed to just one memory allocation and a lookup which is obviously faster, even without running any benchmarks.
```python
if source == '*':
        return ()
    return tuple(source.split('.'))
```
Has a code branch that is completely skipped once it has been used once since it's cached and requires only a lookup instead of memory access, comparison to a string. At least in CPython empty tuples are a singleton so no memory allocation is required.
The other code branch requires us to run two loops and, one that splits the string and allocates a list of strings (so N+1 allocations where N is how many '.' characters are present in the string) and conversion to a tuple which requires us to run the second loop and allocate memory for the tuple.
This code is obviously faster when a serializer is used more than once which is almost always (or even always) the case.
Does it really matter by how much?
If it really does here are the microbenchmarks for ```get_field_label()```:
Without caching:
```
%timeit get_field_label('first_name')
The slowest run took 9.52 times longer than the fastest. This could mean that an intermediate result is being cached 
1000000 loops, best of 3: 300 ns per loop
```
With caching:
```
%timeit get_field_label('first_name')
The slowest run took 10.47 times longer than the fastest. This could mean that an intermediate result is being cached 
100000 loops, best of 3: 2.28 µs per loop
```
Even ~20µs which was the slowest run in the cached version is much much faster than 300 ns.
And here are the microbenchmarks for ```get_source_attributes()```:
Without caching:
```
%timeit get_source_attributes('user.first_name')
The slowest run took 7.01 times longer than the fastest. This could mean that an intermediate result is being cached 
1000000 loops, best of 3: 442 ns per loop
```

With caching:
```
%timeit get_source_attributes('user.first_name')   
The slowest run took 16.55 times longer than the fastest. This could mean that an intermediate result is being cached 
100000 loops, best of 3: 2.31 µs per loop
```

So in total you save almost a milisecond which does make a difference if your goal is to build a responsive API that responds in 100ms or less (or even 50ms or less like github does).
Not to mention that PyPy will JIT the entire function call to something else since it will be able to tell that the value is already cached.

@xordoquy Does that satisfy you?