@xordoquy 2015-10-08
Thanks for the explanations.
Before reading them carefully, I'd just want to put some emphasis on a point.

> So in total you save almost a milisecond which does make a difference if your goal is to build a responsive API that responds in 100ms or less (or even 50ms or less like github does).

Note that DRF goal isn't to perform as well as a finely tuned API. Though better performance is always a good thing DRF politic is to keep the code base as maintainable as possible. We mostly do that on our free time and the fastest it is the better. I'm not assuming anything on your patch or the improvement - just wanted to make this point clear.