@thedrow 2015-10-08
@xordoquy I am aware of that, and I disagree with that mission statement as I'm sure that @tomchristie is aware. APIs created with DRF are not toy-projects. Actual startups and companies use your code.
Maintainable public API is a better goal in my opinion. Not introducing a new utility function that runs faster just because it complicates the code a little bit isn't a tradeoff I'd accept. This is one of the reasons why I don't volunteer for to maintain the project.